# api_spec.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/idphoto":{"post":{"summary":"Idphoto Inference","operationId":"idphoto_inference_idphoto_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_idphoto_inference_idphoto_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/human_matting":{"post":{"summary":"Human Matting Inference","operationId":"human_matting_inference_human_matting_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_human_matting_inference_human_matting_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/add_background":{"post":{"summary":"Photo Add Background","operationId":"photo_add_background_add_background_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_photo_add_background_add_background_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/generate_layout_photos":{"post":{"summary":"Generate Layout Photos","operationId":"generate_layout_photos_generate_layout_photos_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_generate_layout_photos_generate_layout_photos_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/watermark":{"post":{"summary":"Watermark","operationId":"watermark_watermark_post","parameters":[{"name":"size","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Size"}},{"name":"opacity","in":"query","required":false,"schema":{"type":"number","default":0.5,"title":"Opacity"}},{"name":"angle","in":"query","required":false,"schema":{"type":"integer","default":30,"title":"Angle"}},{"name":"color","in":"query","required":false,"schema":{"type":"string","default":"#000000","title":"Color"}},{"name":"space","in":"query","required":false,"schema":{"type":"integer","default":25,"title":"Space"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_watermark_watermark_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/set_kb":{"post":{"summary":"Set Kb","operationId":"set_kb_set_kb_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_set_kb_set_kb_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/idphoto_crop":{"post":{"summary":"Idphoto Crop Inference","operationId":"idphoto_crop_inference_idphoto_crop_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_idphoto_crop_inference_idphoto_crop_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_generate_layout_photos_generate_layout_photos_post":{"properties":{"input_image":{"type":"string","format":"binary","title":"Input Image"},"input_image_base64":{"type":"string","title":"Input Image Base64"},"height":{"type":"integer","title":"Height","default":413},"width":{"type":"integer","title":"Width","default":295},"kb":{"type":"integer","title":"Kb"},"dpi":{"type":"integer","title":"Dpi","default":300}},"type":"object","title":"Body_generate_layout_photos_generate_layout_photos_post"},"Body_human_matting_inference_human_matting_post":{"properties":{"input_image":{"type":"string","format":"binary","title":"Input Image"},"input_image_base64":{"type":"string","title":"Input Image Base64"},"human_matting_model":{"type":"string","title":"Human Matting Model","default":"hivision_modnet"},"dpi":{"type":"integer","title":"Dpi","default":300}},"type":"object","title":"Body_human_matting_inference_human_matting_post"},"Body_idphoto_crop_inference_idphoto_crop_post":{"properties":{"input_image":{"type":"string","format":"binary","title":"Input Image"},"input_image_base64":{"type":"string","title":"Input Image Base64"},"height":{"type":"integer","title":"Height","default":413},"width":{"type":"integer","title":"Width","default":295},"face_detect_model":{"type":"string","title":"Face Detect Model","default":"mtcnn"},"hd":{"type":"boolean","title":"Hd","default":true},"dpi":{"type":"integer","title":"Dpi","default":300},"head_measure_ratio":{"type":"number","title":"Head Measure Ratio","default":0.2},"head_height_ratio":{"type":"number","title":"Head Height Ratio","default":0.45},"top_distance_max":{"type":"number","title":"Top Distance Max","default":0.12},"top_distance_min":{"type":"number","title":"Top Distance Min","default":0.1}},"type":"object","title":"Body_idphoto_crop_inference_idphoto_crop_post"},"Body_idphoto_inference_idphoto_post":{"properties":{"input_image":{"type":"string","format":"binary","title":"Input Image"},"input_image_base64":{"type":"string","title":"Input Image Base64"},"height":{"type":"integer","title":"Height","default":413},"width":{"type":"integer","title":"Width","default":295},"human_matting_model":{"type":"string","title":"Human Matting Model","default":"modnet_photographic_portrait_matting"},"face_detect_model":{"type":"string","title":"Face Detect Model","default":"mtcnn"},"hd":{"type":"boolean","title":"Hd","default":true},"dpi":{"type":"integer","title":"Dpi","default":300},"face_align":{"type":"boolean","title":"Face Align","default":false},"whitening_strength":{"type":"integer","title":"Whitening Strength","default":0},"head_measure_ratio":{"type":"number","title":"Head Measure Ratio","default":0.2},"head_height_ratio":{"type":"number","title":"Head Height Ratio","default":0.45},"top_distance_max":{"type":"number","title":"Top Distance Max","default":0.12},"top_distance_min":{"type":"number","title":"Top Distance Min","default":0.1},"brightness_strength":{"type":"number","title":"Brightness Strength","default":0},"contrast_strength":{"type":"number","title":"Contrast Strength","default":0},"sharpen_strength":{"type":"number","title":"Sharpen Strength","default":0},"saturation_strength":{"type":"number","title":"Saturation Strength","default":0}},"type":"object","title":"Body_idphoto_inference_idphoto_post"},"Body_photo_add_background_add_background_post":{"properties":{"input_image":{"type":"string","format":"binary","title":"Input Image"},"input_image_base64":{"type":"string","title":"Input Image Base64"},"color":{"type":"string","title":"Color","default":"000000"},"kb":{"type":"integer","title":"Kb"},"dpi":{"type":"integer","title":"Dpi","default":300},"render":{"type":"integer","title":"Render","default":0}},"type":"object","title":"Body_photo_add_background_add_background_post"},"Body_set_kb_set_kb_post":{"properties":{"input_image":{"type":"string","format":"binary","title":"Input Image"},"input_image_base64":{"type":"string","title":"Input Image Base64"},"dpi":{"type":"integer","title":"Dpi","default":300},"kb":{"type":"integer","title":"Kb","default":50}},"type":"object","title":"Body_set_kb_set_kb_post"},"Body_watermark_watermark_post":{"properties":{"input_image":{"type":"string","format":"binary","title":"Input Image"},"input_image_base64":{"type":"string","title":"Input Image Base64"},"text":{"type":"string","title":"Text","default":"Hello"},"kb":{"type":"integer","title":"Kb"},"dpi":{"type":"integer","title":"Dpi","default":300}},"type":"object","title":"Body_watermark_watermark_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}